requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
daphne==4.0.0
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.4
djangorestframework==3.14.0
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.12.1
flake8==6.1.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
MarkupSafe==2.1.3
mccabe==0.7.0
msgpack==1.0.7
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
pyOpenSSL==23.3.0
pytest-django==4.8.0
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.31.1
requests==2.31.0
rpds-py==0.13.2
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.30.1
whitenoise==6.6.0
zope.interface==6.1
